pyproject.toml
[tool.poetry]
name = "crossref_journal_minting"
version = "0.1.0"
description = ""
authors = ["Mark Baggett <mbagget1@utk.edu>"]

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
lxml = "^4.9.1"
arrow = "^1.2.3"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
